+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Version 2.0.37
 - Fixed bug for nodeservers that do not pass in name to polyinterface.Interface call.
+- Added PyLogger class as suggested by @firstone
 
 ### Version 2.0.36
 - New logger format with name instead of module